fix: Complete refactoring of Table component documentation page
