+++ docs/src/collections/table/tableSampleFixed.jsx
@@ -1,0 +1,114 @@
+import React from 'react';
+import {
+    Header,
+    Table,
+} from 'react-cm-ui';
+import Highlighter from '../../app/highlighter';
+
+const fixedTableSample = `import React from 'react';
+import { Table } from 'react-cm-ui';
+
+export default class FixedTableSample extends React.Component {
+    render() {
+        return (
+            <Table fixed>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell>Header 1</Table.HeaderCell>
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+        );
+    }
+}`;
+
+function TableSampleFixed() {
+    return (
+        <React.Fragment>
+            <Header anchor="fixed" size="large" style={{ marginTop: '55px' }} sub>
+                Fixed Table
+                <Header.Subheader>
+                    A table column&rsquo;s width can be evenly spaced.
+                </Header.Subheader>
+            </Header>
+
+            <Table fixed>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell>Header 1</Table.HeaderCell>
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>
+                            Lorem ipsum dolor sit amet, consectetur
+                            adipiscing elit, sed do eiusmod tempor
+                            incididunt ut labore et dolore magna aliqua.
+                        </Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>
+                            Lorem ipsum dolor sit amet, consectetur
+                            adipiscing elit, sed do eiusmod tempor
+                            incididunt ut labore et dolore magna aliqua.
+                        </Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>
+                            Lorem ipsum dolor sit amet, consectetur
+                            adipiscing elit, sed do eiusmod tempor
+                            incididunt ut labore et dolore magna aliqua.
+                        </Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+
+            <Highlighter customStyle={{ marginBottom: '44px', marginTop: '44px' }}>
+                {fixedTableSample}
+            </Highlighter>
+        </React.Fragment>
+    );
+}
+
+export default TableSampleFixed;
