+++ docs/src/collections/table/tableSampleFullWidth.jsx
@@ -1,0 +1,105 @@
+import React from 'react';
+import {
+    Header,
+    Table,
+} from 'react-cm-ui';
+import Highlighter from '../../app/highlighter';
+
+const fullWidthTableSample = `import React from 'react';
+import { Table } from 'react-cm-ui';
+
+export default class FullWidthTableSample extends React.Component {
+    render() {
+        return (
+            <Table definition fullWidth>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell />
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+        );
+    }
+}`;
+
+function TableSampleFullWidth() {
+    return (
+        <React.Fragment>
+            <Header anchor="full-width" size="large" style={{ marginTop: '55px' }} sub>
+                Full Width Table
+                <Header.Subheader>
+                    A table, along with
+                    <code>definition</code>
+                    enabled, can have a full width header, filling in the
+                    gap left by the first column.
+                </Header.Subheader>
+            </Header>
+
+            <Table definition fullWidth>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell />
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+
+            <Highlighter customStyle={{ marginBottom: '44px', marginTop: '44px' }}>
+                {fullWidthTableSample}
+            </Highlighter>
+        </React.Fragment>
+    );
+}
+
+export default TableSampleFullWidth;
