+++ docs/src/collections/table/tableSampleStriped.jsx
@@ -1,0 +1,117 @@
+import React from 'react';
+import {
+    Header,
+    Table,
+} from 'react-cm-ui';
+import Highlighter from '../../app/highlighter';
+
+const stripedTableSample = `import React from 'react';
+import { Table } from 'react-cm-ui';
+
+export default class StripedTableSample extends React.Component {
+    render() {
+        return (
+            <Table striped>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell>Header 1</Table.HeaderCell>
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+        );
+    }
+}`;
+
+function TableSampleStriped() {
+    return (
+        <React.Fragment>
+            <Header anchor="striped" size="large" style={{ marginTop: '55px' }} sub>
+                Striped Table
+                <Header.Subheader>
+                    A table&rsquo;s rows can be striped, alternating rows
+                    with slightly darker background color.
+                </Header.Subheader>
+            </Header>
+
+            <Table striped>
+                <Table.Header>
+                    <Table.Row>
+                        <Table.HeaderCell>Header 1</Table.HeaderCell>
+                        <Table.HeaderCell>Header 2</Table.HeaderCell>
+                        <Table.HeaderCell>Header 3</Table.HeaderCell>
+                        <Table.HeaderCell>Header 4</Table.HeaderCell>
+                    </Table.Row>
+                </Table.Header>
+
+                <Table.Body>
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+
+                    <Table.Row>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                        <Table.Cell>cell</Table.Cell>
+                    </Table.Row>
+                </Table.Body>
+            </Table>
+
+            <Highlighter customStyle={{ marginBottom: '44px', marginTop: '44px' }}>
+                {stripedTableSample}
+            </Highlighter>
+        </React.Fragment>
+    );
+}
+
+export default TableSampleStriped;
